# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    DriverProfileInput:
      type: object
      required: [unitNumber, phoneNumber]
      properties:
        unitNumber: { type: string, minLength: 1, maxLength: 20 }
        phoneNumber: { type: string, minLength: 8, maxLength: 40 }
    DriverProfile:
      type: object
      required: [username, unitNumber, phoneNumber]
      properties:
        username: { type: string }
        unitNumber: { type: string }
        phoneNumber: { type: string, description: Normalized phone number. This does not assert phone verification. }
    DriverProfileResponse:
      type: object
      required: [profile]
      properties:
        profile:
          oneOf:
            - $ref: '#/components/schemas/DriverProfile'
            - type: 'null'
    DriverScheduledTrip:
      type: object
      required: [busNumber, officialRunKey, serviceDate, direction, status, scheduledDepartureAt, destinationAddress, stopCount, assignmentSource, displayKeys, keyLegend, scheduleKeysAvailable]
      properties:
        busNumber: { type: string }
        officialRunKey: { type: string }
        serviceDate: { type: string }
        direction: { type: string }
        status:
          type: string
          enum: [ready, running]
        scheduledDepartureAt: { type: string, format: date-time }
        destinationAddress: { type: string }
        stopCount: { type: integer, minimum: 0 }
        assignmentSource:
          type: string
          enum: [dispatch, driver]
        displayKeys:
          type: array
          description: Current rider-facing PDF legend keys for this published run, kept separate from its persisted operational route.
          items: { type: string, enum: [B, G, H, M, P, Q, R, X] }
        keyLegend:
          type: array
          description: Current meanings returned by the centralized official schedule source.
          items:
            type: object
            required: [key, meaning]
            properties:
              key: { type: string, enum: [B, G, H, M, P, Q, R, X] }
              meaning: { type: string }
        scheduleKeysAvailable:
          type: boolean
          description: False when the current official schedule could not be read or no longer contains this run; the persisted trip remains unchanged.
    DriverScheduledTripList:
      type: array
      items: { $ref: '#/components/schemas/DriverScheduledTrip' }
...
paths:
...
  /driver/release-coaches:
    post:
      operationId: releaseDriverCoaches
      tags: [driver]
      security: [{ clerkSession: [] }]
      responses:
        "204":
          description: Releases every coach owned by this driver before calling Clerk signOut
        "401":
          description: Driver sign-in required
  /driver/profile:
    get:
      operationId: getDriverProfile
      tags: [driver]
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: The signed-in driver's own profile, or null when onboarding is incomplete
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverProfileResponse' }
        "401":
          description: Driver sign-in required
    put:
      operationId: updateDriverProfile
      tags: [driver]
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverProfileInput' }
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverProfileResponse' }
        "400":
          description: Invalid profile
        "401":
          description: Driver sign-in required
        "409":
          description: Username already in use (case insensitive)
  /driver/scheduled-trips:
    get:
      operationId: getDriverScheduledTrips
      tags: [driver]
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Active scheduled trips assigned to or selected by the signed-in driver
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverScheduledTripList' }
        "401":
          description: Driver sign-in required
        "403":
          description: Driver access or completed profile required
...